Fixed cyclic module dependency
