+++ src/main/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './utils';
+export * from './lang';
